+++ src/features/total-evaluation/hooks/use-get-portfolio-feedback-data.ts
@@ -1,6 +1,6 @@
 import { useParams } from 'react-router';
 
-import { useGetPortfolioFeedback } from '@/features/total-evaluation/services/use-get-portfolio-feedback';
+import { useGetPortfolioFeedback } from '@/features/feedback/services/use-get-portfolio-feedback';
 
 export const useGetPortfolioFeedbackData = () => {
   const { feedbackId } = useParams();
